feat: server와 handler, dto를 파일로 구분
